Created League model

+++ db/seeds.rb
@@ -7,13 +7,16 @@
 #   ["Action", "Comedy", "Drama", "Horror"].each do |genre_name|
 #     MovieGenre.find_or_create_by!(name: genre_name)
 #   end
-[Game, Tournament, User, Player ].each(&:destroy_all)
 
-Player.create(name: 'Victor', surname: 'Gonzalez', ema_number: '10990053')
-Player.create(name: 'Rosa', surname: 'Melano', ema_number: '22222222')
-Player.create(name: 'Larra', surname: 'Jalano', ema_number: '22222223')
-Player.create(name: 'Aitor', surname: 'Tilla', ema_number: '22222224')
+unless Rails.env.production?
+  [Game, Tournament, User, Player].each(&:destroy_all)
 
-victor = User.create(email_address: 'victor.gf87@gmail.com', password: 'jaja')
+  Player.create(name: 'Victor', surname: 'Gonzalez', ema_number: '10990053')
+  Player.create(name: 'Rosa', surname: 'Melano', ema_number: '22222222')
+  Player.create(name: 'Larra', surname: 'Jalano', ema_number: '22222223')
+  Player.create(name: 'Aitor', surname: 'Tilla', ema_number: '22222224')
 
-Tournament.create(name: 'Sekai Taikai', players: Player.all, creator: victor)
+  victor = User.create(email_address: 'victor.gf87@gmail.com', password: 'jaja')
+
+  Tournament.create(name: 'Sekai Taikai', players: Player.all, creator: victor)
+end
